Started the second project
Filled in the blank spots with the information required on the CSS and HTML pages. linked the CSS pages. set the charset, styled the html, body, body footer, and address. Put style rules for H1 and H2 tags. Set up the nav to look and act fancy. Finished the article rules.